refactor(volume-backup): some update
ref: MANAGER-10569
