+++ packages/manager/modules/pci/src/projects/project/storages/volume-backup/volume-backup.html
@@ -1,0 +1,6 @@
+<div data-ui-view="volumeBackupView">
+
+    <!--View displays-->
+    <div data-ui-view></div>
+    
+</div>
