+++ packages/manager/modules/pci/src/projects/project/storages/volume-backup/volume-backup.module.js
@@ -5,12 +5,11 @@ import '@ovh-ux/ng-translate-async-loader';
 import '@ovh-ux/ng-ovh-cloud-universe-components';
 import '@ovh-ux/ui-kit';
 
-import routing from './volume-backups.routing';
-import component from './volume-backups.component';
+import routing from './volume-backup.routing';
+import component from './volume-backup.component';
 import service from './volume-backup.service';
 
 // import addVolumeBackup from './add';
-// import dashboard from './dashboard';
 import onboarding from './onboarding';
 
 const moduleName = 'ovhManagerPciProjectStoragesVolumeBackup';
