--- packages/manager/modules/pci/src/projects/project/storages/volume-backup/volume-backups.html
@@ -1,8 +1,0 @@
-<div data-ui-view="volumeBackupView">
-    <!--TODO: will be completed on dedicated User Story-->
-
-    <div
-        class="container-fluid w-100 h-100 pt-4 px-3 px-md-5"
-        data-ui-view
-    ></div>
-</div>
